Made several expansions and merges.
Fixed some syntax.
Italicised "Adventure Time" in a couple of places which were missed.
Corrected spelling mistake: Olsen -> Olson
Rearranged media list to reflect canonicity rather than release date.
Corrected a small mistake in the explanation of the X digit.
Reworded a paragraph in "The Beginning of the Multiverse" so it reads better.
Added a paragraph in "The Structure of Existence" giving an example of temporal synchronisation.
Elaborated on Death and the Cosmic Owl in "The Cosmic Entities".
Rearranged "The Concept of Death" to read better.
Added a brief explanation of the elements in "The Formation of the Solar System". In the same section, added a little more speculation about Hunson's origins.
Expanded "Other Worlds" with a lot more detail and explanation.
Changed the date of "The Cosmic Elementals" from 4.5 BYA to 3.5 BYA.
Added more details on Lumps in "The Earthly Elementals".
Added "comet avatar" as an alternative name for Agents of Change in "The Catalyst Comet".
Slightly expanded "The Twin Gems" and added a note.
Mentioned absence of the Lumps Elemental in "Evergreen's Crown and the Green Catalyst Comet".
Expanded "Humans and Vampires".
Switched "Hunson Takes the Nightosphere" and "The First Coming of GOLB and the Death of Margles" to improve flow.
Added a little further explanation in "The First Coming of GOLB and the Death of Margles".
Merged all pre-Lincoln mentions of the penguin form of Orgalorg into one new section called "Orgalorg Through History".
Merged "Sveinn" and "The Middle Ages and Saint Nicholas" into one new section called "The Journey of Evergreen's Crown".
Reworded "The Assassination of Abraham Lincoln" to improve clarity.
Merged the rest of the Orgalorg sections into one new section called "The Further Adventures of Orgalorg".
Slightly expanded "The Human Space Age".
Expanded "Simon and Betty".
Slightly expanded "The Birth of Marceline Abadeer".
Added a list of other pre-war media to "The Fionna and Cake Implication".
Slightly expanded "Prelude to the War".
Expanded "The Great Mushroom War".
Minor tweaks to "Weekly Records in the Wreckage of the World".
Slightly expanded "Simon Meets a Demon" and "The Great Tidal Wave".
Expanded "Marceline and Hunson".
Minor tweaks to "Marceline the Vampire Hunter".
Added note about the apostrophe in "Founders' Island" to the section "The Foundation of the Human Islands".
Moved most of the Bubbline stuff to a new page.
Merged "Flame King" into "The Fire Kingdom" and added a note about his real name.
Expanded "Ice King and Hunson's Collaborative Project".
Corrected the label of "Little Dude" from S05E11 to S05E10.
Slightly expanded "The Birth of Lumpy Space Princess" and "The Birth of Phoebe the Flame Princess".
Added a quote to "Frieda's Flight".
Moved "Gata" to improve flow.
Added quote to "The Time Slips".
Minor tweaks to "Jake's Complicated Past".
Added links to the new media page from "Sources" and "Adventure Time!"
Added BMO Bonanza to the list of present-day stories.
Merged "Cuber's Childhood" into "The Graybles".
Tweaked the dates in the Mnemonoid timeline.